Add UI and dao classes

+++ lab2/src/main/java/omaftiyak/javacourse/lab2/model/Book.java
@@ -9,7 +9,7 @@ public class Book {
 
     public static final String EMPTY_STRING = "";
 
-    private int id;
+    private long id;
     private String bookTitle;
     private String author;
     private int yearPublication;
@@ -31,7 +31,7 @@ public class Book {
         this(book.id, book.yearPublication, book.author, book.bookTitle, book.description, book.language, book.genre);
     }
 
-    public Book(int id, int yearPublication, String author, String bookTitle, String description, String language, String genre) {
+    public Book(long id, int yearPublication, String author, String bookTitle, String description, String language, String genre) {
         this.bookTitle = bookTitle;
         this.author = author;
         this.yearPublication = yearPublication;
@@ -87,11 +87,11 @@ public class Book {
 
     };
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
